Revert "Merge branch 'dev' into BEBIG-216--mission/monthly-batch"
This reverts commit a474afce6e7f6703d036248955ca10fde4d30a83, reversing
changes made to 98b4be514b1528c6118f0c542d0b5048cd9a935e.
