+++ src/main/java/beBig/controller/CommunityController.java
@@ -38,8 +38,7 @@ public class CommunityController {
     public ResponseEntity<PostListResponseDto> list(@RequestParam(value = "limit", defaultValue = "10", required = false) Optional<Integer> limit,
                                                     @RequestParam(value = "offset", defaultValue = "0", required = false) Optional<Integer> offset,
                                                     @RequestParam(value = "category", required = false) Optional<Integer> postCategory,
-                                                    @RequestParam(value = "type", required = false) Optional<Integer> finTypeCode,
-                                                    @RequestParam(value = "sort", required = false) Optional<String> sortType) {
+                                                    @RequestParam(value = "type", required = false) Optional<Integer> finTypeCode) {
         // Optional에서 값이 없을 경우 -1로 처리
         int category = postCategory.orElse(-1);
         log.info("category: " + category);
@@ -48,9 +47,8 @@ public class CommunityController {
         int page = offset.orElse(0);
         log.info("page: " + page);
         int pageSize = limit.orElse(10);
-        String sort = sortType.orElse("newest");
 
-        PostListResponseDto postListResponseDto = communityService.showList(category, type,page,pageSize, sort);
+        PostListResponseDto postListResponseDto = communityService.showList(category, type,page,pageSize);
         if (postListResponseDto.getList() == null || postListResponseDto.getList().isEmpty()) {
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
         }
