+++ src/main/java/beBig/mapper/MissionMapper.java
@@ -91,17 +91,5 @@ public interface MissionMapper {
 //    int countSavingsAccounts(@Param("userId")Long userId);
 
     void updateMonthlyMissionStatus(@Param("personalMonthlyMissionId")int personalMonthlyMissionId, @Param("status") int status);
-
-    long findMonthlyMissionIdByUserId(@Param("userId") long userId);
-
-    int countUserPosts(@Param("thisMonth") int thisMonth, @Param("userId") long userId);
-
-    int countUserLikes(@Param("thisMonth") int thisMonth, @Param("userId") long userId);
-
-    int getMonthlyConsumption(@Param("thisMonth") int thisMonth, @Param("accountNum") String accountNum);
-
-    List<String> getAccountListByUserId(@Param("userId") long userId);
-
-    int getDailyConsumption(@Param("thisMonth") int thisMonth, @Param("day") int day, @Param("accountNum") String accountNum);
 }
