+++ src/main/resources/beBig/mapper/UserMapper.xml
@@ -187,4 +187,5 @@
         from refresh_token
         where user_id = #{userId}
     </select>
+
 </mapper>
